+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/ConfigurationTest.php
@@ -214,7 +214,7 @@ class ConfigurationTest extends \PHPUnit_Framework_TestCase
                 'cache' => 'validator.mapping.cache.symfony',
             ),
             'annotations' => array(
-                'cache' => 'file',
+                'cache' => 'php_array',
                 'file_cache_dir' => '%kernel.cache_dir%/annotations',
                 'debug' => true,
             ),
@@ -274,6 +274,10 @@ class ConfigurationTest extends \PHPUnit_Framework_TestCase
                 'default_redis_provider' => 'redis://localhost',
             ),
             'workflows' => array(),
+            'php_errors' => array(
+                'log' => false,
+                'throw' => true,
+            ),
         );
     }
 }