+++ src/Symfony/Component/Cache/Tests/Adapter/AdapterTestCase.php
@@ -15,6 +15,15 @@ use Cache\IntegrationTests\CachePoolTest;
 
 abstract class AdapterTestCase extends CachePoolTest
 {
+    protected function setUp()
+    {
+        parent::setUp();
+
+        if (!array_key_exists('testDeferredSaveWithoutCommit', $this->skippedTests) && defined('HHVM_VERSION')) {
+            $this->skippedTests['testDeferredSaveWithoutCommit'] = 'Destructors are called late on HHVM.';
+        }
+    }
+
     public function testDefaultLifeTime()
     {
         if (isset($this->skippedTests[__FUNCTION__])) {
@@ -23,18 +32,56 @@ abstract class AdapterTestCase extends CachePoolTest
             return;
         }
 
-        $this->cache = $this->createCachePool(2);
+        $cache = $this->createCachePool(2);
 
-        $item = $this->cache->getItem('key.dlt');
+        $item = $cache->getItem('key.dlt');
         $item->set('value');
-        $this->cache->save($item);
+        $cache->save($item);
         sleep(1);
 
-        $item = $this->cache->getItem('key.dlt');
+        $item = $cache->getItem('key.dlt');
         $this->assertTrue($item->isHit());
 
         sleep(2);
-        $item = $this->cache->getItem('key.dlt');
+        $item = $cache->getItem('key.dlt');
         $this->assertFalse($item->isHit());
+    }
+
+    public function testNotUnserializable()
+    {
+        if (isset($this->skippedTests[__FUNCTION__])) {
+            $this->markTestSkipped($this->skippedTests[__FUNCTION__]);
+
+            return;
+        }
+
+        $cache = $this->createCachePool();
+
+        $item = $cache->getItem('foo');
+        $cache->save($item->set(new NotUnserializable()));
+
+        $item = $cache->getItem('foo');
+        $this->assertFalse($item->isHit());
+
+        foreach ($cache->getItems(array('foo')) as $item) {
+        }
+        $cache->save($item->set(new NotUnserializable()));
+
+        foreach ($cache->getItems(array('foo')) as $item) {
+        }
+        $this->assertFalse($item->isHit());
+    }
+}
+
+class NotUnserializable implements \Serializable
+{
+    public function serialize()
+    {
+        return serialize(123);
+    }
+
+    public function unserialize($ser)
+    {
+        throw new \Exception(__CLASS__);
     }
 }