+++ src/Symfony/Component/VarDumper/Tests/Caster/ExceptionCasterTest.php
@@ -40,12 +40,12 @@ Exception {
   #file: "%sExceptionCasterTest.php"
   #line: 23
   -trace: {
-    %d. %sExceptionCasterTest.php: {
+    %d. %sExceptionCasterTest.php:23: {
       22: {
       23:     return new \Exception('foo');
       24: }
     }
-    %d. %sExceptionCasterTest.php: {
+    %d. %sExceptionCasterTest.php:%d: {
       %d: {
       %d:     $e = $this->getTestException(1);
       %d: 
@@ -57,6 +57,31 @@ Exception {
 EODUMP;
 
         $this->assertDumpMatchesFormat($expectedDump, $e);
+    }
+
+    public function testSeek()
+    {
+        $e = $this->getTestException(2);
+
+        $expectedDump = <<<'EODUMP'
+{
+  %d. %sExceptionCasterTest.php:23: {
+    22: {
+    23:     return new \Exception('foo');
+    24: }
+  }
+  %d. %sExceptionCasterTest.php:%d: {
+    %d: {
+    %d:     $e = $this->getTestException(2);
+    %d: 
+    args: {
+      2
+    }
+  }
+%A
+EODUMP;
+
+        $this->assertStringMatchesFormat($expectedDump, $this->getDump($e, 'trace'));
     }
 
     public function testNoArgs()
@@ -71,12 +96,12 @@ Exception {
   #file: "%sExceptionCasterTest.php"
   #line: 23
   -trace: {
-    %d. %sExceptionCasterTest.php: {
+    %d. %sExceptionCasterTest.php:23: {
       22: {
       23:     return new \Exception('foo');
       24: }
     }
-    %d. %sExceptionCasterTest.php: {
+    %d. %sExceptionCasterTest.php:%d: {
       %d: {
       %d:     $e = $this->getTestException(1);
       %d:     ExceptionCaster::$traceArgs = false;