+++ src/Symfony/Component/Yaml/CHANGELOG.md
@@ -4,11 +4,17 @@ CHANGELOG
 3.2.0
 -----
 
+ * Mappings with a colon that is not followed by a space are deprecated and
+   will lead to a `ParseException` in Symfony 4.0.
+
  * Added support for parsing PHP constants:
 
    ```php
    Yaml::parse('!php/const:PHP_INT_MAX', Yaml::PARSE_CONSTANT);
    ```
+
+ * Support for silently ignoring duplicate keys in YAML has been deprecated and
+   will lead to a `ParseException` in Symfony 4.0.
 
 3.1.0
 -----